desqer v0.0.1 Desqer.AppointmentController

Endpoints for appointments.

Functions

create(conn, map)

Creates appointment or appointments.

Endpoint example:

POST /appointments

Request headers example:

authorization: Bearer eyJhbGciOiJIUzUxMiI5c...

Params example by customer:

{
  "appointment": {
    "service_id": "80e9863b-61bc-45be-9f0e-9aebfb2dcec6",
    "starts_at": "2017-03-10 15:30:00",
    "ends_at": "2017-03-10 16:00:00",
    "notes": "Hair massage needed"
  }
}

Params example by professional:

{
  "appointment": {
    "service_id": "80e9863b-61bc-45be-9f0e-9aebfb2dcec6",
    "name": "Haircut",
    "description": "The perfect hair style for your type of face",
    "price": 4900,
    "starts_at": "2017-03-10 15:30:00",
    "ends_at": "2017-03-10 16:00:00",
    "notes": "Hair massage bonus"
  },
  "user_ids": [
    "69f06156-7fc2-4c7f-abde-fc20204627e1"
  ]
}

Params details:

price in cents

Success response code 200 example for customer:

{
  "data": {
    "appointment": {
      "id": "2050ea22-a273-4bef-93e9-ce9df0e73ddc",
      "user_id": "69f06156-7fc2-4c7f-abde-fc20204627e1",
      "service_id": "9c1c0135-9cb7-41d6-9991-9c8042d31bc8",
      "name": "Haircut",
      "description": "The perfect hair style for your type of face",
      "price": 4900,
      "starts_at": "2017-03-10T15:30:00",
      "ends_at": "2017-03-10T16:00:00",
      "notes": "Hair massage needed",
      "status": "active",
      "owned": nil
    }
  }
}

Success response code 200 example for professional:

{
  "data": {
    "appointments": [
      {
        "id": "2050ea22-a273-4bef-93e9-ce9df0e73ddc",
        "user_id": "69f06156-7fc2-4c7f-abde-fc20204627e1",
        "service_id": "9c1c0135-9cb7-41d6-9991-9c8042d31bc8",
        "name": "Haircut",
        "description": "The perfect hair style for your type of face",
        "price": 4900,
        "starts_at": "2017-03-10T15:30:00",
        "ends_at": "2017-03-10T16:00:00",
        "notes": "Hair massage bonus",
        "status": "active",
        "owned": nil
      }
    ]
  }
}

Error response code 422 example:

{
  "errors": {
    "starts_at": ["can't be blank"],
    "ends_at": ["can't be blank"]
  }
}

Error response code 401 example:

{
  "errors": {
    "detail": "Resource not authorized"
  }
}

index(conn, params)

List appointments.

Endpoint example:

GET /appointments

Request headers example:

authorization: Bearer eyJhbGciOiJIUzUxMiI5c...

Success response code 200 example:

{
  "data": [
    {
      "id": "2050ea22-a273-4bef-93e9-ce9df0e73ddc",
      "user_id": "69f06156-7fc2-4c7f-abde-fc20204627e1",
      "service_id": "9c1c0135-9cb7-41d6-9991-9c8042d31bc8",
      "name": "Haircut",
      "description": "The perfect hair style for your type of face",
      "price": 4900,
      "starts_at": "2017-03-10T15:30:00",
      "ends_at": "2017-03-10T16:00:00",
      "notes": "Hair massage bonus",
      "status": "active",
      "owned": true
    }
  ]
}

Error response code 401 example:

{
  "errors": {
    "detail": "Resource not authorized"
  }
}

update(conn, map)

Updates appointment.

Endpoint example:

PUT /appointments/2050ea22-a273-4bef-93e9-ce9df0e73ddc

Request headers example:

authorization: Bearer eyJhbGciOiJIUzUxMiI5c...

Params example:

{
  "appointment": {
    "starts_at": "2017-03-09 13:00",
    "ends_at": "2017-03-09 13:30",
    "notes": "Hair massage bonus and others",
    "status": "canceled"
  }
}

Params details:

status in [active, canceled]

Success response code 200 example:

{
  "data": {
    "appointment": {
      "id": "2050ea22-a273-4bef-93e9-ce9df0e73ddc",
      "user_id": "69f06156-7fc2-4c7f-abde-fc20204627e1",
      "service_id": "9c1c0135-9cb7-41d6-9991-9c8042d31bc8",
      "name": "Haircut",
      "description": "The perfect hair style for your type of face",
      "price": 4900,
      "starts_at": "2017-03-09 13:00",
      "ends_at": "2017-03-09 13:30",
      "notes": "Hair massage bonus and others",
      "status": "canceled",
      "owned": nil
    }
  }
}

Error response code 422 example:

{
  "errors": {
    "status": ["is invalid"]
  }
}

Error response code 401 example:

{
  "errors": {
    "detail": "Resource not authorized"
  }
}